backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.0
cryptography==42.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
pillow==10.3.0
langchain==0.1.20
langchain-community==0.0.38
tavily-python==0.3.3
aiofiles==23.2.1
httpx==0.27.0
bcrypt==4.3.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
